# Optimize Jacobi eigensolver (WIP)