--- a/library/src/include/lib_host_helpers.hpp
+++ b/library/src/include/lib_host_helpers.hpp
@@ -54,16 +54,6 @@ __device__ __host__ inline int64_t
     return j * lda + i * inca;
 }
 
-inline int64_t idx2D(const int32_t i, const int32_t j, const int32_t lda)
-{
-    return j * static_cast<int64_t>(lda) + i;
-}
-
-inline int64_t idx2D(const int32_t i, const int32_t j, const int32_t inca, const int32_t lda)
-{
-    return j * static_cast<int64_t>(lda) + i * static_cast<int64_t>(inca);
-}
-
 inline int32_t ceil(const int32_t n, const int32_t nb)
 {
     return ((n - 1) / nb) + 1;