This PR updates the kernels for Jacobi eigensolver (syevj) to improve the performance several kernels such as  offd_rotate() that apply the block rotation matrix "J" to off-diagonal blocks.  The offd_rotate()  kernel may account for nearly 90% of the execution time on the GPU. 

The PR converts some of the kernels to use the grid-stride or "for-loop" coding style so that the launch configuration can be modified. The offd_rotate() kernel is modified to  store the sub-matrix blocks of matrix "A" and rotation matrix "J" in LDS shared memory. The launch configuration is modified to encourate the reuse of matrix "J" already in LDS to update several sub-matrix blocks of matrix "A".

On the gfx1030, there is about a 35% performance improvement in rocsolver-bench  for n=4000,  with eigenvectors requested in single and double precisions.

For example

rocsolver-bench -f syevj  -n  4000 --evect V --precision s --perf 1 --iters 1

original time 9.9 sec, new time 6.5 sec

rocsolver-bench -f syevj  -n  4000 --evect V --precision d --perf 1 --iters 1

original time 19.1 sec, new time 13.1 sec




Note that the algorithmic block size is currently set to be 32 and this prevents the matrix J to be loaded in LDS for double_complex type.  Further improvements might be possible if the kernels  can be updated to accept a different algorithmic block size.